+++ Craftsman/Builders/Tests/FunctionalTests/PutEntityTestBuilder.cs
@@ -67,7 +67,7 @@ public class {Path.GetFileNameWithoutExtension(classPath.FullClassPath)} : TestB
         var clientAuth = isProtected ? @$"
 
         var user = await AddNewSuperAdmin();
-        _client.AddAuth(user.Identifier);" : "";
+        FactoryClient.AddAuth(user.Identifier);" : "";
 
         return $@"[Test]
     public async Task {testName}()
@@ -79,7 +79,7 @@ public class {Path.GetFileNameWithoutExtension(classPath.FullClassPath)} : TestB
 
         // Act
         var route = ApiRoutes.{FileNames.GetApiRouteClass(entity.Plural)}.Put.Replace(ApiRoutes.{FileNames.GetApiRouteClass(entity.Plural)}.{pkName}, {fakeEntityVariableName}.{pkName}.ToString());
-        var result = await _client.PutJsonRequestAsync(route, {fakeDtoVariableName});
+        var result = await FactoryClient.PutJsonRequestAsync(route, {fakeDtoVariableName});
 
         // Assert
         result.StatusCode.Should().Be(HttpStatusCode.NoContent);
@@ -107,7 +107,7 @@ public class {Path.GetFileNameWithoutExtension(classPath.FullClassPath)} : TestB
 
         // Act
         var route = ApiRoutes.{FileNames.GetApiRouteClass(entity.Plural)}.Put.Replace(ApiRoutes.{FileNames.GetApiRouteClass(entity.Plural)}.{pkName}, {fakeEntityVariableName}.{pkName}.ToString());
-        var result = await _client.PutJsonRequestAsync(route, {fakeDtoVariableName});
+        var result = await FactoryClient.PutJsonRequestAsync(route, {fakeDtoVariableName});
 
         // Assert
         result.StatusCode.Should().Be(HttpStatusCode.Unauthorized);
@@ -130,13 +130,13 @@ public class {Path.GetFileNameWithoutExtension(classPath.FullClassPath)} : TestB
         // Arrange
         var {fakeEntityVariableName} = {fakeEntity}.Generate(new {fakeCreationDto}().Generate());
         var {fakeDtoVariableName} = new {fakeUpdateDto} {{ }}.Generate();
-        _client.AddAuth();
+        FactoryClient.AddAuth();
 
         await InsertAsync({fakeEntityVariableName});
 
         // Act
         var route = ApiRoutes.{FileNames.GetApiRouteClass(entity.Plural)}.Put.Replace(ApiRoutes.{FileNames.GetApiRouteClass(entity.Plural)}.{pkName}, {fakeEntityVariableName}.{pkName}.ToString());
-        var result = await _client.PutJsonRequestAsync(route, {fakeDtoVariableName});
+        var result = await FactoryClient.PutJsonRequestAsync(route, {fakeDtoVariableName});
 
         // Assert
         result.StatusCode.Should().Be(HttpStatusCode.Forbidden);