+++ Craftsman/Builders/Tests/Utilities/WebAppFactoryBuilder.cs
@@ -13,18 +13,17 @@ public class WebAppFactoryBuilder
         _utilities = utilities;
     }
 
-    public void CreateWebAppFactory(string solutionDirectory, string projectName, string dbContextName, bool addJwtAuthentication)
+    public void CreateWebAppFactory(string testDirectory, string projectName, bool addJwtAuthentication)
     {
-        var classPath = ClassPathHelper.FunctionalTestProjectRootClassPath(solutionDirectory, $"{FileNames.GetWebHostFactoryName()}.cs", projectName);
-        var fileText = GetWebAppFactoryFileText(classPath, dbContextName, solutionDirectory, projectName, addJwtAuthentication);
+        var classPath = ClassPathHelper.FunctionalTestProjectRootClassPath(testDirectory, $"{FileNames.GetWebHostFactoryName()}.cs", projectName);
+        var fileText = GetWebAppFactoryFileText(classPath, testDirectory, projectName, addJwtAuthentication);
         _utilities.CreateFile(classPath, fileText);
     }
 
-    private static string GetWebAppFactoryFileText(ClassPath classPath, string dbContextName, string solutionDirectory, string projectBaseName, bool addJwtAuthentication)
+    private static string GetWebAppFactoryFileText(ClassPath classPath, string testDirectory, string projectBaseName, bool addJwtAuthentication)
     {
-        var webApiClassPath = ClassPathHelper.WebApiProjectRootClassPath(solutionDirectory, "", projectBaseName);
-        var contextClassPath = ClassPathHelper.DbContextClassPath(solutionDirectory, "", projectBaseName);
-        var utilsClassPath = ClassPathHelper.WebApiResourcesClassPath(solutionDirectory, "", projectBaseName);
+        var webApiClassPath = ClassPathHelper.WebApiProjectRootClassPath(testDirectory, "", projectBaseName);
+        var utilsClassPath = ClassPathHelper.WebApiResourcesClassPath(testDirectory, "", projectBaseName);
 
         var authUsing = addJwtAuthentication ? $@"
 using WebMotions.Fake.Authentication.JwtBearer;" : "";
@@ -36,50 +35,39 @@ using WebMotions.Fake.Authentication.JwtBearer;" : "";
                     options.DefaultAuthenticateScheme = FakeJwtBearerDefaults.AuthenticationScheme;
                     options.DefaultChallengeScheme = FakeJwtBearerDefaults.AuthenticationScheme;
                 }}).AddFakeJwtBearer();
+
 " : "";
 
         return @$"
 namespace {classPath.ClassNamespace};
 
-using {contextClassPath.ClassNamespace};
 using {utilsClassPath.ClassNamespace};
 using {webApiClassPath.ClassNamespace};{authUsing}
 using Mapster;
 using MapsterMapper;
 using Microsoft.AspNetCore.Mvc.Testing;
-using Microsoft.EntityFrameworkCore;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using System.Reflection;
+using Microsoft.Extensions.Logging;
 
 public class {Path.GetFileNameWithoutExtension(classPath.FullClassPath)} : WebApplicationFactory<Program>
 {{
     protected override IHost CreateHost(IHostBuilder builder)
     {{
         builder.UseEnvironment(Consts.Testing.FunctionalTestingEnvName);
+        builder.ConfigureLogging(logging =>
+        {{
+            logging.ClearProviders();
+        }});
 
         builder.ConfigureServices(services =>
         {{{authRegistration}
-            var provider = services.BuildServiceProvider();
-
             var typeAdapterConfig = TypeAdapterConfig.GlobalSettings;
             typeAdapterConfig.Scan(Assembly.GetExecutingAssembly());
             var mapperConfig = new Mapper(typeAdapterConfig);
             services.AddSingleton<IMapper>(mapperConfig);
-
-            services.AddDbContext<{dbContextName}>(options =>
-            {{
-                options.UseInMemoryDatabase(""InMemoryDbForTesting"");
-                options.UseInternalServiceProvider(provider);
-            }});
-
-            var sp = services.BuildServiceProvider();
-
-            using var scope = sp.CreateScope();
-            var scopedServices = scope.ServiceProvider;
-            var db = scopedServices.GetRequiredService<{dbContextName}>();
-            db.Database.EnsureCreated();
         }});
         
         return base.CreateHost(builder);