+++ CHANGELOG.md
@@ -52,6 +52,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 * New `Email` Value Object. Used on `User` if you want to see it in use
 
+* Functional tests use a docker db
+
 
 ### Updated